Move TshdClient to createClient.ts
