+++ web/packages/teleterm/src/types.ts
@@ -24,8 +24,10 @@ import { Logger, LoggerService } from 'teleterm/services/logger/types';
 import { FileStorage } from 'teleterm/services/fileStorage';
 import { MainProcessClient, RuntimeSettings } from 'teleterm/mainProcess/types';
 import { PtyServiceClient } from 'teleterm/services/pty';
-import { TshdClient } from 'teleterm/services/tshd/types';
-import { VnetServiceClient } from 'teleterm/services/tshd/createClient';
+import {
+  VnetServiceClient,
+  TshdClient,
+} from 'teleterm/services/tshd';
 
 export type {
   Logger,
