+++ web/packages/teleterm/src/ui/services/headlessAuthn/headlessAuthnService.ts
@@ -21,13 +21,14 @@ import { MainProcessClient } from 'teleterm/types';
 import { ModalsService } from 'teleterm/ui/services/modals';
 import { ConfigService } from 'teleterm/services/config';
 
+import type { TshdClient, CloneableAbortSignal } from 'teleterm/services/tshd';
 import type * as types from 'teleterm/services/tshd/types';
 
 export class HeadlessAuthenticationService {
   constructor(
     private mainProcessClient: MainProcessClient,
     private modalsService: ModalsService,
-    private tshClient: types.TshdClient,
+    private tshClient: TshdClient,
     private configService: ConfigService
   ) {}
 
@@ -60,7 +61,7 @@ export class HeadlessAuthenticationService {
 
   async updateHeadlessAuthenticationState(
     params: types.UpdateHeadlessAuthenticationStateRequest,
-    abortSignal: types.CloneableAbortSignal
+    abortSignal: CloneableAbortSignal
   ): Promise<void> {
     await this.tshClient.updateHeadlessAuthenticationState(params, {
       abort: abortSignal,
