+++ web/packages/teleterm/src/ui/services/resources/resourcesService.test.ts
@@ -30,6 +30,7 @@ import {
   ResourcesService,
 } from './resourcesService';
 
+import type { TshdClient } from 'teleterm/services/tshd';
 import type * as tsh from 'teleterm/services/tshd/types';
 
 describe('getServerByHostname', () => {
@@ -37,7 +38,7 @@ describe('getServerByHostname', () => {
   const getServerByHostnameTests: Array<
     {
       name: string;
-      getServersMockedValue: ReturnType<tsh.TshdClient['getServers']>;
+      getServersMockedValue: ReturnType<TshdClient['getServers']>;
     } & (
       | { expectedServer: tsh.Server; expectedErr?: never }
       | { expectedErr: any; expectedServer?: never }
@@ -74,10 +75,10 @@ describe('getServerByHostname', () => {
   test.each(getServerByHostnameTests)(
     '$name',
     async ({ getServersMockedValue, expectedServer, expectedErr }) => {
-      const tshClient: Partial<tsh.TshdClient> = {
+      const tshClient: Partial<TshdClient> = {
         getServers: jest.fn().mockResolvedValueOnce(getServersMockedValue),
       };
-      const service = new ResourcesService(tshClient as tsh.TshdClient);
+      const service = new ResourcesService(tshClient as TshdClient);
 
       const promise = service.getServerByHostname('/clusters/bar', 'foo');
 
@@ -110,7 +111,7 @@ describe('searchResources', () => {
     const kube = makeKube();
     const app = makeApp();
 
-    const tshClient: Partial<tsh.TshdClient> = {
+    const tshClient: Partial<TshdClient> = {
       getServers: jest.fn().mockResolvedValueOnce(
         new MockedUnaryCall({
           agents: [server],
@@ -140,7 +141,7 @@ describe('searchResources', () => {
         })
       ),
     };
-    const service = new ResourcesService(tshClient as tsh.TshdClient);
+    const service = new ResourcesService(tshClient as TshdClient);
 
     const searchResults = await service.searchResources({
       clusterUri: '/clusters/foo',
@@ -172,7 +173,7 @@ describe('searchResources', () => {
 
   it('returns a single item if a filter is supplied', async () => {
     const server = makeServer();
-    const tshClient: Partial<tsh.TshdClient> = {
+    const tshClient: Partial<TshdClient> = {
       getServers: jest.fn().mockResolvedValueOnce(
         new MockedUnaryCall({
           agents: [server],
@@ -181,7 +182,7 @@ describe('searchResources', () => {
         })
       ),
     };
-    const service = new ResourcesService(tshClient as tsh.TshdClient);
+    const service = new ResourcesService(tshClient as TshdClient);
 
     const searchResults = await service.searchResources({
       clusterUri: '/clusters/foo',
@@ -200,13 +201,13 @@ describe('searchResources', () => {
 
   it('returns a custom error pointing at resource kind and cluster when an underlying promise gets rejected', async () => {
     const expectedCause = new Error('oops');
-    const tshClient: Partial<tsh.TshdClient> = {
+    const tshClient: Partial<TshdClient> = {
       getServers: jest.fn().mockRejectedValueOnce(expectedCause),
       getDatabases: jest.fn().mockRejectedValueOnce(expectedCause),
       getKubes: jest.fn().mockRejectedValueOnce(expectedCause),
       getApps: jest.fn().mockRejectedValueOnce(expectedCause),
     };
-    const service = new ResourcesService(tshClient as tsh.TshdClient);
+    const service = new ResourcesService(tshClient as TshdClient);
 
     const searchResults = await service.searchResources({
       clusterUri: '/clusters/foo',
