+++ web/packages/teleterm/src/ui/types.ts
@@ -37,8 +37,7 @@ import { UsageService } from 'teleterm/ui/services/usage';
 import { ConfigService } from 'teleterm/services/config';
 import { ConnectMyComputerService } from 'teleterm/ui/services/connectMyComputer';
 import { HeadlessAuthenticationService } from 'teleterm/ui/services/headlessAuthn/headlessAuthnService';
-import { TshdClient } from 'teleterm/services/tshd/types';
-import { VnetServiceClient } from 'teleterm/services/tshd/createClient';
+import { TshdClient, VnetServiceClient } from 'teleterm/services/tshd';
 
 export interface IAppContext {
   clustersService: ClustersService;
